ticket:1727 - Petascope collects datetime process requests
Summary: Syncing for petascope data process requests handlers.

Test Plan: Run test case services (system test).

Reviewers: dmisev

Reviewed By: dmisev

Differential Revision: http://codereview.rasdaman.org/D825

+++ applications/petascope/petascope_core/src/main/java/petascope/util/DatabaseUtil.java
@@ -147,10 +147,18 @@ public class DatabaseUtil {
         try {
             connection = dataSource.getConnection();
             
-            statement = connection.createStatement();
-            // just delete the lock before Liquibase starts
-            String selectQuery = "delete from " + LIQUIBASE_LOCK_TABLE;
-            statement.executeUpdate(selectQuery);
+            DatabaseMetaData databaseMetadata = connection.getMetaData();
+            
+            // check if liquibase lock table exists
+            ResultSet tables = databaseMetadata.getTables(null, null, LIQUIBASE_LOCK_TABLE, null);
+            
+            if (tables.next()) {
+                // Only clean the lock if this table exists
+                statement = connection.createStatement();
+                // just delete the lock before Liquibase starts
+                String selectQuery = "delete from " + LIQUIBASE_LOCK_TABLE;
+                statement.executeUpdate(selectQuery);
+            }
         } catch (SQLException ex) {
             throw new PetascopeException(ExceptionCode.InternalSqlError, 
                     "Cannot delete lock of liquibase for petascope database "
